top_state managed by home_alone_base; remove strat_boss
Change HomeAlone strategies to WinterTou and ShoulderTou, since we will likely
start having dynamic tariffs with no time of use rate as well.
