+++ gw_spaceheat/actors/home_alone/winter_using_strat.py
